增加物流单号、评价;修改部分结构接口

+++ src/main/java/com/takefree/controller/TakeApplicationController.java
@@ -3,6 +3,7 @@ package com.takefree.controller;
 import com.takefree.common.Exception.SimpleHttpException;
 import com.takefree.common.annotation.Authorization;
 import com.takefree.common.config.Constants;
+import com.takefree.common.entry.JsonObjectList;
 import com.takefree.common.entry.JsonSimpleObject;
 import com.takefree.common.entry.Token;
 import com.takefree.common.util.JsonObjectUtils;
@@ -12,19 +13,22 @@ import com.takefree.dto.model.TakeApplicationDTO;
 import com.takefree.enums.ApplyStatusEnum;
 import com.takefree.enums.ShareStatusEnum;
 import com.takefree.pojo.model.ShareLike;
+import com.takefree.pojo.model.TakeApplication;
 import com.takefree.service.TakeApplicationService;
 import com.takefree.service.ShareService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.Date;
+import java.util.List;
 
 /**
  * Created by gaoxiang on 2017/11/4.
  */
 @Controller
-public class ShareApplyController {
+public class TakeApplicationController {
     @Autowired
     private ShareService shareService;
 
@@ -34,7 +38,7 @@ public class ShareApplyController {
     @RequestMapping(value = "/apply",method = RequestMethod.POST)
     @ResponseBody
     @Authorization
-    public JsonSimpleObject<ShareLike> createApply(@RequestAttribute(Constants.TAKEFREE_TOKEN) Token token, @RequestBody com.takefree.pojo.model.TakeApplication takeApplication) throws Exception{
+    public JsonSimpleObject<ShareLike> createApply(@RequestAttribute(Constants.TAKEFREE_TOKEN) Token token,@Valid @RequestBody TakeApplication takeApplication) throws Exception{
         ShareDTO shareInfo=shareService.getShareInfoById(takeApplication.getShareId());
 
         if(shareInfo==null){
@@ -50,9 +54,9 @@ public class ShareApplyController {
         }
 
         takeApplication.setApplicantId(token.getUserDTO().getId());
-        takeApplication.setOwnerId(shareInfo.getOwnerId());
+        takeApplication.setOwnerId(shareInfo.getOwnerId()); //建议删掉
         takeApplication.setApplyTime(new Date());
-        takeApplication.setStatus(ApplyStatusEnum.INIT.getCode());
+        //takeApplication.setStatus(ApplyStatusEnum.INIT.getCode());
         /**
          * TODO...
          * address采用引用个方法有bug，修改之后被级联修改
@@ -67,11 +71,30 @@ public class ShareApplyController {
     @ResponseBody
     @Authorization
     public JsonSimpleObject deleteApply(@RequestAttribute(Constants.TAKEFREE_TOKEN) Token token, @PathVariable("id") Long takeApplicationId) throws Exception{
-        com.takefree.pojo.model.TakeApplication apply= takeApplicationService.getById(takeApplicationId);
+        TakeApplicationDTO apply= takeApplicationService.getById(takeApplicationId);
+        if(apply==null){
+            throw new SimpleHttpException(HttpStatus.NOT_FOUND, "无此申请");
+        }
         if(!apply.getApplicantId().equals(token.getUserDTO().getId())){
-            throw new SimpleHttpException(HttpStatus.UNAUTHORIZED, "无权限");
+            throw new SimpleHttpException(HttpStatus.FORBIDDEN, "Unauthorized");
         }
         takeApplicationService.deleteById(takeApplicationId);
+        return JsonObjectUtils.buildSimpleObjectSuccess(null);
+    }
+
+    @RequestMapping(value = "/apply/{id}",method = RequestMethod.PUT)
+    @ResponseBody
+    @Authorization
+    public JsonSimpleObject updateApply(@RequestAttribute(Constants.TAKEFREE_TOKEN) Token token, @PathVariable("id") Long takeApplicationId,@RequestBody TakeApplication takeApplication) throws Exception{
+        TakeApplicationDTO apply= takeApplicationService.getById(takeApplicationId);
+        if(apply==null){
+            throw new SimpleHttpException(HttpStatus.NOT_FOUND, "无此申请");
+        }
+        if(!apply.getApplicantId().equals(token.getUserDTO().getId())){
+            throw new SimpleHttpException(HttpStatus.FORBIDDEN, "Unauthorized");
+        }
+        takeApplication.setId(takeApplicationId);
+        takeApplicationService.updateById(takeApplication);
         return JsonObjectUtils.buildSimpleObjectSuccess(null);
     }
 
@@ -79,9 +102,12 @@ public class ShareApplyController {
     @ResponseBody
     @Authorization
     public JsonSimpleObject<TakeApplicationDTO> getApply(@RequestAttribute(Constants.TAKEFREE_TOKEN) Token token, @PathVariable("id") Long takeApplicationId) throws Exception{
-        com.takefree.pojo.model.TakeApplication apply= takeApplicationService.getById(takeApplicationId);
+        TakeApplicationDTO apply= takeApplicationService.getById(takeApplicationId);
+        if(apply==null){
+            throw new SimpleHttpException(HttpStatus.NOT_FOUND, "无此订单");
+        }
         if(!apply.getApplicantId().equals(token.getUserDTO().getId())){
-            throw new SimpleHttpException(HttpStatus.UNAUTHORIZED, "无权限");
+            throw new SimpleHttpException(HttpStatus.UNAUTHORIZED, "Unauthorized");
         }
         return JsonObjectUtils.buildSimpleObjectSuccess(apply);
     }
@@ -98,12 +124,14 @@ public class ShareApplyController {
      */
     @RequestMapping(value = "/apply",method = RequestMethod.GET)
     @ResponseBody
-    public JsonSimpleObject<TakeApplicationDTO> getApplys(@RequestAttribute(value=Constants.TAKEFREE_TOKEN,required = false) Token token, Integer pageNo, Integer pageSize, Long applyUserId, Long shareId, Integer status) throws Exception{
-        com.takefree.pojo.model.TakeApplication apply= takeApplicationService.getApplys(pageNo, pageSize, shareId, null, applyUserId, status);
-        if(!apply.getApplicantId().equals(token.getUserDTO().getId())){
-            throw new SimpleHttpException(HttpStatus.UNAUTHORIZED, "无权限");
+    public JsonObjectList<TakeApplicationDTO> getApplys(@RequestAttribute(value=Constants.TAKEFREE_TOKEN,required = false) Token token, Integer pageNo, Integer pageSize, Long applyUserId, Long shareId, Integer status) throws Exception{
+        if(applyUserId!=null){
+            if(token==null||!token.getUserDTO().getId().equals(applyUserId)){
+                throw new SimpleHttpException(HttpStatus.UNAUTHORIZED, "Unauthorized");
+            }
         }
-        return JsonObjectUtils.buildSimpleObjectSuccess(apply);
-    }
 
+        List<TakeApplicationDTO> apply= takeApplicationService.getApplys(pageNo, pageSize, shareId, null, applyUserId, status);
+        return JsonObjectUtils.buildListSuccess(apply);
+    }
 }
